# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cashback/me:
    get:
      operationId: getMyCashback
      tags: [cashback]
      summary: Get the signed-in customer's cashback balance and history
      responses:
        "200":
          description: Cashback account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyCashback"
        "401":
          description: Sign-in required
  /admin/cashback/pending:
    get:
      operationId: listPendingCashback
      tags: [cashback]
      summary: List cashback credits awaiting manual approval
      responses:
        "200":
          description: Pending cashback transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCashbackTransaction"
        "401":
          description: Admin sign-in required
  /referral/me:
    get:
      operationId: getMyReferral
      tags: [cashback]
      summary: Get the signed-in customer's referral code and successful referral count
      responses:
        "200":
          description: Referral account
          content:
            application/json:
              schema:
                type: object
                required: [referralCode, referralCount]
                properties:
                  referralCode: { type: string }
                  referralCount: { type: number }
        "401": { description: Sign-in required }
...
  /admin/referrals:
    get:
      operationId: listAdminReferrals
      tags: [admin, cashback]
      summary: List referral performance and rewards
      responses:
        "200":
          description: Referral performance list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminReferral"
        "401": { description: Admin sign-in required }
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List customers and cashback balances
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCustomer"
        "401": { description: Admin sign-in required }
  /admin/users/{customerId}/cashback:
    post:
      operationId: adjustUserCashback
      tags: [admin, cashback]
      summary: Add or deduct cashback for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustUserCashbackInput"
      responses:
        "201": { description: Cashback adjusted }
        "409": { description: Deduction exceeds available balance }
        "401": { description: Admin sign-in required }
  /admin/cashback/{id}/approve:
    post:
      operationId: approveCashback
      tags: [cashback]
      summary: Approve a pending cashback credit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Approved cashback transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCashbackTransaction"
        "404":
          description: Pending cashback transaction not found
        "401":
          description: Admin sign-in required
...
components:
...
  schemas:
...
    CashbackBalance:
      type: object
      required: [currency, pending, available]
      properties:
        currency:
          type: string
          enum: [EGP, USD]
        pending:
          type: number
        available:
          type: number
    CashbackTransaction:
      type: object
      required: [id, orderId, type, status, currency, amount, createdAt, approvedAt]
      properties:
        id:
          type: number
        orderId:
          type: ["number", "null"]
        orderNumber:
          oneOf:
            - type: string
            - type: number
          nullable: true
        type:
          type: string
          enum: [credit, debit]
        status:
          type: string
          enum: [pending, available, redeemed, voided, reversed]
        currency:
          type: string
          enum: [EGP, USD]
        amount:
          type: number
        createdAt:
          type: string
        approvedAt:
          type: [string, "null"]
    AdjustUserCashbackInput:
      type: object
      required: [amount, currency, operation]
      properties:
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          enum: [EGP, USD]
        operation:
          type: string
          enum: [add, deduct]
    MyCashback:
      type: object
      required: [balances, transactions]
      properties:
        balances:
          type: array
          items:
            $ref: "#/components/schemas/CashbackBalance"
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/CashbackTransaction"
    AdminCashbackTransaction:
      allOf:
        - $ref: "#/components/schemas/CashbackTransaction"
        - type: object
          required: [customerName, customerEmail]
          properties:
            customerName:
              type: string
            customerEmail:
              type: string
...
    AdminCustomer:
      type: object
      required: [customerId, name, email, referralCode, orderCount, balances]
      properties:
        customerId:
          type: string
        name:
          type: string
        email:
          type: string
        referralCode:
          type: ["string", "null"]
        orderCount:
          type: number
        balances:
          type: array
          items:
            type: object
            required: [currency, pending, available]
            properties:
              currency:
                type: string
                enum: [EGP, USD]
              pending:
                type: number
              available:
                type: number
...
    AdminReferral:
      type: object
      required: [referralCode, referrerCustomerId, referrerName, referrerEmail, referredCustomers, successfulOrders, pendingRewardEgp, availableRewardEgp, lastActivityAt]
      properties:
        referralCode: { type: string }
        referrerCustomerId: { type: string }
        referrerName: { type: string }
        referrerEmail: { type: string }
        referredCustomers: { type: number }
        successfulOrders: { type: number }
        pendingRewardEgp: { type: number }
        availableRewardEgp: { type: number }
        lastActivityAt: { type: ["string", "null"] }
...