# lib/api-zod/src/generated/types/adminReferral.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminReferral {
  referralCode: string;
  referrerCustomerId: string;
  referrerName: string;
  referrerEmail: string;
  referredCustomers: number;
  successfulOrders: number;
  pendingRewardEgp: number;
  availableRewardEgp: number;
  /** @nullable */
  lastActivityAt: string | null;
}